requirements.txt
streamlit
opencv-python
numpy
joblib
Pillow
scikit-learn
google-generativeai
